fix letters filter and search
