+++ src/app/main/content/mainContent/letters/letters.tsx
@@ -4,8 +4,6 @@ import styles from './letters.module.css';
 import { Letter } from './letter/letter';
 import { LetterType } from '../../../../types/types';
 
-const MAX_LETTERS = 30;
-
 interface IProps {
   letters: LetterType[];
   checkedLetterIds: { [id: string]: boolean };
@@ -15,7 +13,6 @@ interface IProps {
   isDark: boolean;
   removeAddAnimation: (id: string) => void;
   removeLetter: (id: string) => void;
-  setVisibility: (id: string, value: boolean) => void;
 }
 
 export class Letters extends Component {
@@ -24,44 +21,31 @@ export class Letters extends Component {
     this.props = props;
   }
 
-  private searchPredicate = (text: string[]) => {
-    return text.find((value: string) => {
-      return value.includes(this.props.searchText);
-    });
-  };
-
   public readonly props: IProps;
 
   public render() {
-    let visible = 0;
     return (
       <div className={styles.letters}>
         {this.props.letters.map((letter: LetterType) => {
-          if (visible < MAX_LETTERS && this.searchPredicate(letter.text)) {
-            visible += 1;
-            this.props.setVisibility(letter.id, true);
-            return (
-              <Letter
-                key={letter.id}
-                id={letter.id}
-                text={letter.text}
-                authorAbbr={letter.authorAbbr}
-                author={letter.author}
-                subject={letter.subject}
-                date={letter.date}
-                isChecked={this.props.checkedLetterIds[letter.id]}
-                onCheckboxChange={this.props.onCheckboxChange}
-                openLetter={this.props.openLetter}
-                isDark={this.props.isDark}
-                hasAddAnimation={letter.hasAddAnimation}
-                hasDeleteAnimation={letter.hasDeleteAnimation}
-                removeAddAnimation={this.props.removeAddAnimation}
-                removeLetter={this.props.removeLetter}
-              />
-            );
-          }
-          this.props.setVisibility(letter.id, false);
-          return null;
+          return (
+            <Letter
+              key={letter.id}
+              id={letter.id}
+              text={letter.text}
+              authorAbbr={letter.authorAbbr}
+              author={letter.author}
+              subject={letter.subject}
+              date={letter.date}
+              isChecked={this.props.checkedLetterIds[letter.id]}
+              onCheckboxChange={this.props.onCheckboxChange}
+              openLetter={this.props.openLetter}
+              isDark={this.props.isDark}
+              hasAddAnimation={letter.hasAddAnimation}
+              hasDeleteAnimation={letter.hasDeleteAnimation}
+              removeAddAnimation={this.props.removeAddAnimation}
+              removeLetter={this.props.removeLetter}
+            />
+          );
         })}
       </div>
     );
