+++ src/app/main/content/mainContent/mainContent.tsx
@@ -13,7 +13,6 @@ interface IProps {
   isDark: boolean;
   removeAddAnimation: (id: string) => void;
   removeLetter: (id: string) => void;
-  setVisibility: (id: string, value: boolean) => void;
 }
 
 interface IState {
@@ -69,7 +68,6 @@ export class MainContent extends Component {
           isDark={this.props.isDark}
           removeAddAnimation={this.props.removeAddAnimation}
           removeLetter={this.props.removeLetter}
-          setVisibility={this.props.setVisibility}
         />
       </div>
     );
